requirements.txt
torch~=2.6.0
pandas~=2.2.3
kagglehub~=0.3.11
setuptools~=75.8.2
numpy~=2.2.4
torchvision~=0.21.0
pillow~=10.4.0
hydra-core~=1.3.2
mlflow~=2.21.3
mlflow-skinny~=2.21.3
omegaconf~=2.3.0
tqdm~=4.67.1
PyYAML~=6.0.2
matplotlib~=3.10.1
moviepy~=2.1.2
